ConstantValueManager: Move to new module
